requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Geospatial Analysis
geopandas>=0.13.0
rasterio>=1.3.0
xarray>=2023.1.0
dask>=2023.1.0
pyproj>=3.5.0
shapely>=2.0.0
folium>=0.14.0

# Machine Learning
lightgbm>=4.0.0
xgboost>=2.0.0
optuna>=3.1.0
shap>=0.42.0

# Statistical Analysis
statsmodels>=0.14.0
econml>=0.14.0
causalml>=0.13.0
linearmodels>=4.27

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
bokeh>=3.1.0

# Climate Data
cdsapi>=0.5.1
netCDF4>=1.6.0
h5py>=3.8.0

# Web Scraping & APIs
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.8.0

# Utilities
tqdm>=4.65.0
joblib>=1.2.0
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0

# Development
jupyter>=1.0.0
ipykernel>=6.20.0
black>=23.0.0
flake8>=6.0.0
pytest>=7.2.0

# Optional (for Google Earth Engine)
earthengine-api>=0.1.350

# Additional data acquisition dependencies
wbdata>=0.3.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0